% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Title}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{number of times code will run, more iterations leads to better results}

\item{x}{a vector of values}

\item{fun}{function that can look for desired statistic}

\item{alpha}{probability that the statistic does not lie within calculated range}

\item{...}{parameters for the histogram output}
}
\value{
a histogram and a list of statistics that might be of interest
}
\description{
Title
}
\examples{
myboot2(iter=1000, x = (1:10), col="#F34567")
}
