% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{mycltb}
\usage{
mycltb(n, iter, p = 0.5)
}
\arguments{
\item{n}{number of items in the sample}

\item{iter}{how many samples you want to create}

\item{p}{probability of an event being a success}
}
\value{
A histogram of the sample means, and a vector of those sample means
}
\description{
mycltb
}
\examples{
mycltb(n=10,iter=1000,p=0.3)
}
