% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{myhyper}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{number of iteration, higher produces more accurate results}

\item{N}{total number of items in the collection}

\item{r}{total number of "win" conditions in the collection}

\item{n}{number of items in the sample}
}
\value{
a barplot showing the distribution of the hypergeometric sim and a table of the probabilities
}
\description{
myhyper
}
\examples{
myhyper(iter=1000,n=19, N=20,r=12)
}
