% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function of the loglikelihood}

\item{x}{a vector of samples}

\item{param}{a vector of parameter values}

\item{...}{other parameters for plot}
}
\value{
a list of important information and a graph
}
\description{
mymaxlik
}
\examples{
\dontrun{mymaxlik(lfun=logbin, x = c(3,3,4,3,4,5,5,4),
 param = seq(0,1,length=1000),main="Binomial, n=20")}
}
