% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title}
\usage{
myncurve(a, mu, sigma, color = "Blue")
}
\arguments{
\item{a}{value x is less than or equal to}

\item{mu}{mean of the curve}

\item{sigma}{standard deviation of the curve}

\item{color}{color of shaded region}
}
\value{
a plot of the distribution
}
\description{
Title
}
\examples{
myncurve(2,0,1,color="Red")
}
