% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{Title}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{T value}

\item{xmax}{something I'm sure}

\item{n}{number of items in the sample}

\item{alpha}{1-alpha is the confidence}
}
\value{
a list with the bounds of the rejection region q and the pvalue pvalue
}
\description{
Title
}
\examples{
mypvalue(1.326252, n=30, alpha = 0.05)
}
