% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Title}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats on the flight}

\item{gamma}{accaptable probability of overbooking}

\item{p}{probability someone shows up to their flight}
}
\value{
Two graphs showing the ideal number of tickets sold, one continuous and one discrete in math
         and a list of the input variables and the outputs
}
\description{
Title
}
\examples{
ntickets(N=200,gamma=0.02,p=0.95)
}
